version_4.0

+++ src/main/java/com/example/axbit/controller/AbstractControllerImpl.java
@@ -1,6 +1,7 @@
 package com.example.axbit.controller;
 
 import com.example.axbit.exception.EntityNotFoundException;
+import com.example.axbit.exception.NotDeleteException;
 import com.example.axbit.model.AbstractEntity;
 import com.example.axbit.service.AbstractService;
 import org.apache.logging.log4j.LogManager;
@@ -15,7 +16,7 @@ import java.util.List;
 public abstract class AbstractControllerImpl<T extends AbstractEntity, S extends AbstractService<T>>
         implements AbstractController<T> {
     private final S service;
-    private static final Logger LOGGER = LogManager.getLogger(AbstractService.class);
+    private final Logger logger = LogManager.getLogger(AbstractService.class);
 
     public AbstractControllerImpl(S service) {
         this.service = service;
@@ -23,30 +24,38 @@ public abstract class AbstractControllerImpl<T extends AbstractEntity, S extends
 
     @Override
     public ResponseEntity<List<T>> getAllEntity() {
-            List<T> entities = new ArrayList<>(service.getAllEntity());
-            if (entities.isEmpty()) {
-                LOGGER.debug("Entities exist db");
-                throw new EntityNotFoundException("Entities not found");
-            }
-            LOGGER.debug("Entities retrieved from db");
-            return new ResponseEntity<>(entities, HttpStatus.OK);
+        List<T> entities = new ArrayList<>(service.getAllEntity());
+        if (entities.isEmpty()) {
+            logger.debug("Entities exist db");
+            throw new EntityNotFoundException("Entities not found");
+        }
+        logger.debug("Entities retrieved from db");
+        return new ResponseEntity<>(entities, HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<T> getEntityById(@PathVariable Long id) {
-        LOGGER.debug("Input id: " + id);
+        logger.debug("Input id: " + id);
         T entity = service.getEntityById(id);
         if (entity == null) {
-            LOGGER.debug("Entity with id " + id + " exist db");
+            logger.debug("Entity with id " + id + " exist db");
             throw new EntityNotFoundException("Entity id not found: " + id);
         }
-        LOGGER.debug("Entity with id " + id + " retrieved from db");
+        logger.debug("Entity with id " + id + " retrieved from db");
         return new ResponseEntity<>(entity, HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<HttpStatus> deleteEntityById(@PathVariable Long id) {
+        logger.debug("Input delete Entity id: " + id);
+        AbstractEntity entity = service.getEntityById(id);
+        logger.debug("Retrieved from db: " + entity);
+        if (entity == null) {
+            logger.debug("Entity not deleted");
+            throw new NotDeleteException("Entity not found and not deleted");
+        }
         service.deleteEntityById(id);
+        logger.debug("Entity deleted");
         return new ResponseEntity<>(HttpStatus.OK);
     }
 }
